update linux multi build script

+++ raylib/static/build_linux_multi.sh
@@ -1,11 +1,16 @@
 #!/usr/bin/env bash
+if ! command -v pyenv &> /dev/null
+then
+    echo "Pyenv required: https://github.com/pyenv/pyenv-installer"
+    exit
+fi
 rm *linux-gnu.so
-pyenv install -s 3.9.0
-pyenv global 3.9.0
+pyenv install -s 3.9.1
+pyenv global 3.9.1
 pip3 install cffi
 python build_linux.py
-pyenv install -s 3.8.6
-pyenv global 3.8.6
+pyenv install -s 3.8.7
+pyenv global 3.8.7
 pip3 install cffi
 python build_linux.py
 pyenv install -s 3.7.9
